# src/api/swagger.yml
swagger: "2.0"
info:
  title: Monocular
  version: 0.0.1
produces:
  - application/json
consumes:
  - application/json
schemes:
  - http
paths:
  /healthz:
    get:
      operationId: healthz
      summary: "health check"
      responses:
        200:
          description: healthy
        default:
          description: unexpected error
  /v1/charts/{repo}/{chartName}:
    get:
      operationId: getChart
      summary: "get chart by repo/name"
      tags:
        - charts
      parameters:
        - name: repo
          in: path
          type: string
          required: true
        - name: chartName
          in: path
          type: string
          required: true
      responses:
        200:
          description: chart response
          schema:
            $ref: "#/definitions/resourceData"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
  /v1/charts/{repo}/{chartName}/versions:
    get:
      operationId: getChartVersions
      summary: "get all chart versions by repo/name"
      tags:
        - charts
      parameters:
        - name: repo
          in: path
          type: string
          required: true
        - name: chartName
          in: path
          type: string
          required: true
      responses:
        200:
          description: chart response
          schema:
            $ref: "#/definitions/resourceData"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
  /v1/charts/{repo}/{chartName}/versions/{version}:
    get:
      operationId: getChartVersion
      summary: "get specific chart version by repo/name"
      tags:
        - charts
      parameters:
        - name: repo
          in: path
          type: string
          required: true
        - name: chartName
          in: path
          type: string
          required: true
        - name: version
          in: path
          type: string
          required: true
      responses:
        200:
          description: charts response
          schema:
            $ref: "#/definitions/resourceArrayData"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
  /v1/charts:
    get:
      operationId: getAllCharts
      summary: "get all charts from all repos"
      tags:
        - charts
      responses:
        200:
          description: charts response
          schema:
            $ref: "#/definitions/resourceArrayData"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"

  /v1/charts/search:
    get:
      operationId: searchCharts
      summary: "search charts"
      tags:
        - charts
      parameters:
       - $ref: "#/parameters/chartNameParam"
      responses:
        200:
          description: repositories response
          schema:
            $ref: "#/definitions/resourceArrayData"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"


  /v1/charts/{repo}:
    get:
      operationId: getChartsInRepo
      summary: "get all charts by repo"
      tags:
        - charts
      parameters:
        - name: repo
          in: path
          type: string
          required: true
      responses:
        200:
          description: charts by repo response
          schema:
            type: object
            properties:
              data:
                $ref: "#/definitions/resource"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
  /v1/repos:
    get:
      operationId: getAllRepos
      summary: "get all repositories enabled in the backend"
      tags:
        - repositories
      responses:
        200:
          description: repositories response
          schema:
            $ref: "#/definitions/resourceArrayData"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
  /v1/releases:
    get:
      operationId: getAllReleases
      summary: "get all Helm releases"
      tags:
        - releases
      responses:
        200:
          description: Get a list of releases
          schema:
            $ref: "#/definitions/resourceArrayData"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
    post:
      operationId: createRelease
      summary: "Install a Helm chart version creating a release"
      tags:
        - releases
      parameters:
        - $ref: '#/parameters/createReleaseParams'
      responses:
        201:
          description: Release created
          schema:
            $ref: "#/definitions/resourceData"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
  /v1/releases/{releaseName}:
    get:
      operationId: getRelease
      summary: get a release from your cluster
      tags:
        - releases
      parameters:
        - name: releaseName
          in: path
          type: string
          required: true
      responses:
        200:
          description: release information
          schema:
            $ref: "#/definitions/resourceData"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
    delete:
      operationId: deleteRelease
      summary: delete a release from your cluster
      tags:
        - releases
      parameters:
        - name: releaseName
          in: path
          type: string
          required: true
      responses:
        200:
          description: release deleted
          schema:
            $ref: "#/definitions/resourceData"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/error"
parameters:
  chartNameParam:
    name: name
    in: query
    type: string
    description: chart name substring pattern match
    required: true
  createReleaseParams:
    name: data
    description: Information related with the new Helm release.
    in: body
    type: object
    schema:
      properties:
        releaseName:
          type: string
          description: Name of the new release
        chartId:
          type: string
          description: "Chart identifier, i.e: stable/wordpress"
        chartVersion:
          type: string
          description: Chart Version
        dryRun:
          type: boolean
          description: Simulate an install
        namespace:
          type: string
          description: Namespace to deploy the chart into
      required:
        - chartId
        - chartVersion
definitions:
  resource: # a data type that describes a json-api-like resource
    type: object
    required:
      - type
      - id
      - attributes
    properties:
      type:
        type: string
        minLength: 1
      id:
        type: string
        minLength: 1
      links:
        type: object
      attributes:
        type: object
      relationships:
        type: object
  resourceData: # a data type that includes a "data" property that contains a json-api-like resource
    type: object
    required:
      - data
    properties:
      data:
        $ref: "#/definitions/resource"
  resourceArrayData: # a data type that includes a "data" property that contains an array of json-api-like resources
    type: object
    required:
      - data
    properties:
      data:
        type: array
        items:
          $ref: "#/definitions/resource"
  maintainer: # a data type that describes chart maintainer metadata
    type: object
    required:
      - email
      - name
    properties:
      email:
        type: string
        minLength: 1
      name:
        type: string
        minLength: 1
  chartPackage: # a data type that closely conforms to a repo's index.yaml chart entry format
    type: object
    required:
      - name
      - created
      - digest
      - description
      - home
      - maintainers
      - sources
      - urls
      - version
    properties:
      repo:
        type: string
      name:
        type: string
        minLength: 1
      created:
        type: string
        minLength: 1
      digest:
        type: string
        minLength: 1
      description:
        type: string
        minLength: 1
      icon:
        type: string
        minLength: 1
      keywords:
        type: array
        items:
          type: string
      maintainers:
        type: array
        items:
          $ref: "#/definitions/maintainer"
      version:
        type: string
        minLength: 1
      home:
        type: string
        minLength: 1
      sources:
        type: array
        items:
          type: string
      urls:
        type: array
        items:
          type: string
  chart: # a data type that describes a "chart"
    type: object
    required:
      - name
      - description
      - home
      - repo
      - sources
      - maintainers
    properties:
      name:
        type: string
        minLength: 1
      description:
        type: string
        minLength: 1
      repo:
        type: object
        $ref: "#/definitions/repo"
      repoURL:
        type: string
        minLenth: 1
      home:
        type: string
        minLength: 1
      sources:
        type: array
        items:
          type: string
      maintainers:
        type: array
        items:
          $ref: "#/definitions/maintainer"
      keywords:
        type: array
        items:
          type: string
  chartVersion: # a data type that describes version metadata about a "chart version"
    type: object
    required:
      - created
      - digest
      - urls
      - version
      - readme
    properties:
      created:
        type: string
        minLength: 1
      readme:
        type: string
        minLength: 1
      digest:
        type: string
        minLength: 1
      urls:
        type: array
        items:
          type: string
      icons:
        type: array
        items:
          $ref: "#/definitions/icon"
      version:
        type: string
        minLength: 1
  resourceLink: # a data type that describes the "canonical" API endpoint to a resource
    type: object
    required:
      - self
    properties:
      self:
        type: string
        minLength: 1
  chartAsRelationship: # a data type that describes a "chart" as a json-api-like "relationship"
    type: object
    required:
      - links
      - data
    properties:
      links:
        $ref: "#/definitions/resourceLink"
      data:
        $ref: "#/definitions/chart"
  chartVersionAsRelationship: # a data type that describes a "chartVersion" as a json-api-like "relationship"
    type: object
    required:
      - links
      - data
    properties:
      links:
        $ref: "#/definitions/resourceLink"
      data:
        $ref: "#/definitions/chartVersion"
  chartRelationship: # a data type that describes a json-api-like "relationship" to a "chart" type
    type: object
    required:
      - chart
    properties:
      chart:
        $ref: "#/definitions/chartAsRelationship"
  latestChartVersionRelationship: # a data type that describes a json-api-like "relationship" to a "chartVersion" type
    type: object
    required:
      - latestChartVersion
    properties:
      latestChartVersion:
        $ref: "#/definitions/chartVersionAsRelationship"
  error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int64
      message:
        type: string
  repo:
    type: object
    required:
      - name
      - URL
    properties:
      name:
        type: string
        minLength: 1
      URL:
        type: string
        minLength: 1
      source:
        type: string
  icon:
    type: object
    required:
      - name
      - path
    properties:
      name:
        type: string
      path:
        type: string
  release:
    type: object
    required:
      - name
      - status
      - updated
      - chartName
      - chartVersion
      - namespace
    properties:
      name:
        type: string
      status:
        type: string
      updated:
        type: string
      chartName:
        type: string
      chartVersion:
        type: string
      namespace:
        type: string
  releaseExtended:
    type: object
    required:
      - name
      - status
      - updated
      - chartName
      - chartVersion
      - namespace
      - resources
      - notes
      - manifest
    properties:
      name:
        type: string
      status:
        type: string
      updated:
        type: string
      chartName:
        type: string
      chartVersion:
        type: string
      namespace:
        type: string
      name:
        type: string
      resources:
        type: string
      notes:
        type: string
      manifest:
        type: string